+++ AngularClient/src/main/core/entities/element-cell.ts
@@ -27,16 +27,22 @@ export class ElementCell extends EntityBase {
   decimalValueUpdated = new Subject<number>();
 
   private fields: {
+    allRoundsIncome: number,
     currentUserDecimalValue: number,
     income: number,
     decimalValue: number,
     decimalValuePercentage: number,
   } = {
+      allRoundsIncome: 0,
       currentUserDecimalValue: 0,
       income: 0,
       decimalValue: 0,
       decimalValuePercentage: 0,
     };
+
+  allRoundsIncome() {
+    return this.fields.allRoundsIncome;
+  }
 
   currentUserDecimalValue() {
     return this.fields.currentUserDecimalValue;
@@ -46,7 +52,11 @@ export class ElementCell extends EntityBase {
     return this.fields.income;
   }
 
-  initialize(): boolean {
+  increaseAllRoundsIncome() {
+    this.fields.allRoundsIncome += this.income();
+  }
+
+  initialize() {
     if (!super.initialize()) return false;
 
     // Other users'
@@ -99,6 +109,10 @@ export class ElementCell extends EntityBase {
         ? this.currentUserDecimalValue()
         : this.otherUsersDecimalValueTotal
       : this.otherUsersDecimalValueTotal + this.currentUserDecimalValue();
+  }
+
+  resetAllRoundsIncome() {
+    this.fields.allRoundsIncome = 0;
   }
 
   setCurrentUserDecimalValue() {