+++ AngularClient/src/main/core/entities/element-item.ts
@@ -14,17 +14,20 @@ export class ElementItem extends EntityBase {
   ElementCellSet: ElementCell[];
   ParentCellSet: ElementCell[];
 
+  allRoundsIncomeUpdated = new Subject<number>();
   incomeUpdated = new Subject<number>();
 
   // Client-side
   private fields: {
+    allRoundsIncome: number,
     income: number,
   } = {
+      allRoundsIncome: 0,
       income: 0,
     };
 
-  ratingCells() {
-    return this.getRatingCells(this);
+  allRoundsIncome() {
+    return this.fields.allRoundsIncome;
   }
 
   elementCell(fieldName: string): ElementCell {
@@ -63,6 +66,12 @@ export class ElementItem extends EntityBase {
     return "average";
   }
 
+  increaseAllRoundsIncome() {
+    this.fields.allRoundsIncome += this.income();
+
+    this.allRoundsIncomeUpdated.next(this.fields.allRoundsIncome);
+  }
+
   initialize(): boolean {
     if (!super.initialize()) return false;
 
@@ -72,6 +81,16 @@ export class ElementItem extends EntityBase {
     });
 
     return true;
+  }
+
+  ratingCells() {
+    return this.getRatingCells(this);
+  }
+
+  resetAllRoundsIncome() {
+    this.fields.allRoundsIncome = 0;
+
+    this.allRoundsIncomeUpdated.next(this.fields.allRoundsIncome);
   }
 
   setIncome(): void {