+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ### Changelog
 
+**0.86.6**
+
+* Timer related calculation fixes
+* Project - Sorting elements by using "Sort Order" field
+* Minor updates
+
 **0.86.5**
 
 * Project Editor - Add Fields